added move method for GameService added getByCoordinate renamed fields in Coordinates
