+++ src/main/java/io/game/st3/domain/Coordinate.java
@@ -1,0 +1,15 @@
+package io.game.st3.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Coordinate {
+
+    private int row;
+    private int col;
+
+}
