+++ src/main/java/io/game/st3/domain/enums/GameStatus.java
@@ -1,0 +1,8 @@
+package io.game.st3.domain.enums;
+
+public enum GameStatus {
+
+    NEW,
+    IN_PROGRESS,
+    WON
+}
