+++ src/main/java/io/game/st3/domain/enums/XO.java
@@ -1,0 +1,7 @@
+package io.game.st3.domain.enums;
+
+public enum XO {
+    X,
+    O,
+    EMPTY
+}
