--- src/main/java/io/game/superttc/domain/Coordinate.java
@@ -1,15 +1,0 @@
-package io.game.superttc.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Coordinate {
-
-    private int row;
-    private int col;
-
-}
