--- src/main/java/io/game/superttc/domain/Game.java
@@ -1,24 +1,0 @@
-package io.game.superttc.domain;
-
-import java.util.List;
-import java.util.UUID;
-
-import io.game.superttc.domain.enums.GameStatus;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Game {
-    private UUID uuid;
-    private Player player1;
-    private Player player2;
-    private GameStatus gameStatus;
-    private boolean firstPlayerMove = true;
-    private List<Coordinate> validPlayBoards;
-    private Board gameBoard;
-}
