--- src/main/java/io/game/superttc/domain/enums/GameStatus.java
@@ -1,8 +1,0 @@
-package io.game.superttc.domain.enums;
-
-public enum GameStatus {
-
-    NEW,
-    IN_PROGRESS,
-    WON
-}
