--- src/main/java/io/game/superttc/domain/enums/XO.java
@@ -1,7 +1,0 @@
-package io.game.superttc.domain.enums;
-
-public enum XO {
-    X,
-    O,
-    EMPTY
-}
