+++ src/test/java/io/game/st3/ApplicationTests.java
@@ -1,0 +1,13 @@
+package io.game.st3;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
