+++ src/test/java/io/game/st3/utils/TestConstants.java
@@ -1,0 +1,12 @@
+package io.game.st3.utils;
+
+import java.util.UUID;
+
+public class TestConstants {
+
+    public static final int BOARD_SIZE = 3;
+    public static final String FIRST_PLAYER_NAME = "first_player";
+    public static final String SECOND_PLAYER_NAME = "second_player";
+    public static final UUID ID = UUID.fromString("5167ee98-7eb3-493d-ac88-16528018f41e");
+
+}
