--- src/test/java/io/game/superttc/ApplicationTests.java
@@ -1,13 +1,0 @@
-package io.game.superttc;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
